add url 'style' parameter to allow multiple them rendering in a single instance
