+++ mapsforgesrv/src/main/java/com/telemaxx/mapsforgesrv/MapsforgeStyleConfig.java
@@ -1,0 +1,189 @@
+package com.telemaxx.mapsforgesrv;
+
+import java.io.File;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class MapsforgeStyleConfig extends PropertiesParser{
+	
+	private File themeFile = null;
+	private String styleName = null;
+	private String[] themeFileOverlays = null;
+	private String themeFileStyle = null;
+	private float deviceScale;
+	private float userScale;
+	private float textScale;
+	private float symbolScale;
+	private float lineScale;
+	protected double[] hillShadingArguments;
+	protected String hillShadingAlgorithm = null;
+	protected double hillShadingMagnitude;
+	private int blackValue;
+	private double gammaValue;
+	
+	private final static double DEFAULTGAMMA = 1.;
+	private final static int DEFAULTBLACK = 0;
+	
+	public final double[] HILLSHADINGSIMPLEDEFAULT = { 0.1, 0.666 };
+	public final double HILLSHADINGDIFDUSELIGHTDEFAULT = 50;
+	protected final static double DEFAULTHSMAGNITUDE = 1.;
+	
+	private final static float DEFAULTDEVICESCALE = 1.0f;
+	private final static float DEFAULTUSERSCALE = 1.0f;
+	private final static float DEFAULTTEXTSCALE = 1.0f;
+	private final static float DEFAULTSYMBOLSCALE = 1.0f;
+	private final static float DEFAULTLINESCALE = 1.0f;
+	
+	private final static Logger logger = LoggerFactory.getLogger(MapsforgeStyleConfig.class);
+	
+	public MapsforgeStyleConfig(String styleName, File styleFile) throws Exception {
+		this.styleName = styleName;
+		readConfig(styleFile);
+		initConfig();
+	}
+	
+	private void parseThemeFile() throws Exception {
+		String configValue = "themefile";
+		String configString = retrieveConfigValue(configValue);
+		String msgHeader = "Theme";
+		if (configString == null) {
+			themeFile = new File("OSMARENDER");
+			logger.info(parsePadMsg(msgHeader + " " + FILE) + ": default [OSMARENDER]"); //$NON-NLS-1$
+		} else if (configString.trim().equals("OSMARENDER")) {
+			themeFile = new File("OSMARENDER");
+			logger.info(parsePadMsg(msgHeader + " " + FILE) + ": defined [OSMARENDER]"); //$NON-NLS-1$
+		} else if (configString.trim().equals("DEFAULT")) {
+			themeFile = new File("DEFAULT");
+			logger.info(parsePadMsg(msgHeader + " " + FILE) + ": defined [DEFAULT]"); //$NON-NLS-1$
+		} else {
+			themeFile = parseFile(configValue, FILE, false, msgHeader, "OSMARENDER");
+		}
+	}
+	
+	private void parseThemeOverlays() throws Exception {
+		String msgHeader = parsePadMsg("Theme overlay(s)"); //$NON-NLS-1$
+		String optionValue = retrieveConfigValue("overlays"); //$NON-NLS-1$
+		if (optionValue != null)
+			themeFileOverlays = StringUtils.stripAll(optionValue.trim().split(",")); //$NON-NLS-1$
+		if (themeFileOverlays != null && themeFileOverlays.length != 0) {
+			logger.info(msgHeader + ": defined [{" + String.join(",", themeFileOverlays) + "}]"); //$NON-NLS-1$
+		} else {
+			logger.info(msgHeader + ": default [undefined]"); //$NON-NLS-1$
+		}
+	}
+
+	
+	
+	private void initConfig() throws Exception {
+		logger.info("################## STYLE '"+styleName+"' CONFIG ##################");
+		parseThemeFile();
+		themeFileStyle = parseString(null, "style", null, "Theme style"); //$NON-NLS-1$ //$NON-NLS-2$
+		parseThemeOverlays();
+		parseHillShading();
+		hillShadingMagnitude = (double) parseNumber(DEFAULTHSMAGNITUDE, "hillshading-magnitude", 0., 4., "Hillshading magnitude",false); //$NON-NLS-1$ //$NON-NLS-2$
+		deviceScale = (float) parseNumber(DEFAULTDEVICESCALE, "device-scale", 0., null, "Device scale factor",true); //$NON-NLS-1$ //$NON-NLS-2$
+		userScale = (float) parseNumber(DEFAULTUSERSCALE, "user-scale", 0., null, "User scale factor",true); //$NON-NLS-1$ //$NON-NLS-2$
+		textScale = (float) parseNumber(DEFAULTTEXTSCALE, "text-scale", 0., null, "Text scale factor",true); //$NON-NLS-1$ //$NON-NLS-2$
+		symbolScale = (float) parseNumber(DEFAULTSYMBOLSCALE, "symbol-scale", 0., null, "Symbol scale factor",true); //$NON-NLS-1$ //$NON-NLS-2$
+		lineScale = (float) parseNumber(DEFAULTLINESCALE, "line-scale", 0., null, "Line scale factor",true); //$NON-NLS-1$ //$NON-NLS-2$
+		parseHillShading();
+		hillShadingMagnitude = (double) parseNumber(DEFAULTHSMAGNITUDE, "hillshading-magnitude", 0., 4., "Hillshading magnitude",false); //$NON-NLS-1$ //$NON-NLS-2$
+		blackValue = (int) parseNumber(DEFAULTBLACK, "contrast-stretch", 0, 254, "Contrast stretch",false); //$NON-NLS-1$ //$NON-NLS-2$
+		gammaValue = (double) parseNumber(DEFAULTGAMMA, "gamma-correction", 0., null, "Gamma correction",true); //$NON-NLS-1$ //$NON-NLS-2$
+		
+	}
+	
+	private void parseHillShading() throws Exception {
+		String msgHeader = parsePadMsg("HillShading algorithm");
+		hillShadingArguments = null;
+		String hillShadingOption = retrieveConfigValue("hillshading-algorithm"); //$NON-NLS-1$
+		if (hillShadingOption != null) {
+			hillShadingOption = hillShadingOption.trim();
+			Pattern P = Pattern.compile("(simple)(?:\\((\\d+\\.?\\d*|\\d*\\.?\\d+),(\\d+\\.?\\d*|\\d*\\.?\\d+)\\))?|(diffuselight)(?:\\((\\d+\\.?\\d*|\\d*\\.?\\d+)\\))?");
+			Matcher m = P.matcher(hillShadingOption);
+			if (m.matches()) {
+				if (m.group(1) != null) {
+					hillShadingAlgorithm = new String(m.group(1)); // ShadingAlgorithm = simple
+					hillShadingArguments = new double[2];
+					if (m.group(2) != null) {
+						hillShadingArguments[0] = Double.parseDouble(m.group(2));
+						hillShadingArguments[1] = Double.parseDouble(m.group(3));
+					} else { // default values
+						hillShadingArguments[0] = HILLSHADINGSIMPLEDEFAULT[0];
+						hillShadingArguments[1] = HILLSHADINGSIMPLEDEFAULT[1];
+					}
+					logger.info(msgHeader + ": defined [" + hillShadingAlgorithm + "(" + hillShadingArguments[0] + "," //$NON-NLS-3$
+							+ hillShadingArguments[1] + ")]");
+				} else {
+					hillShadingAlgorithm = new String(m.group(4)); // ShadingAlgorithm = diffuselight
+					hillShadingArguments = new double[1];
+					if (m.group(5) != null) {
+						hillShadingArguments[0] = Double.parseDouble(m.group(5));
+					} else { // default value
+						hillShadingArguments[0] = HILLSHADINGDIFDUSELIGHTDEFAULT;
+					}
+					logger.info(msgHeader + ": defined [" + hillShadingAlgorithm + "(" + hillShadingArguments[0] + ")]"); //$NON-NLS-1$
+				}
+			} else {
+				parseError(msgHeader, "'" + hillShadingOption + "' invalid", "undefined");
+			}
+		}
+	}
+	
+	public double getHillShadingMagnitude() {
+		return this.hillShadingMagnitude;
+	}
+
+	public String getHillShadingAlgorithm() {
+		return this.hillShadingAlgorithm;
+	}
+
+	public double[] getHillShadingArguments() {
+		return this.hillShadingArguments;
+	}
+	
+	public String getThemeFileStyle() {
+		return this.themeFileStyle;
+	}
+
+	public String[] getThemeFileOverlays() {
+		return this.themeFileOverlays;
+	}
+	
+	public File getThemeFile() {
+		return this.themeFile;
+	}
+	
+	public float getDeviceScale() {
+		return this.deviceScale;
+	}
+	
+	public float getUserScale() {
+		return this.userScale;
+	}
+		
+	public float getTextScale() {
+		return this.textScale;
+	}
+	
+	public float getSymbolScale() {
+		return this.symbolScale;
+	}
+
+	public float getLineScale() {
+		return this.lineScale;
+	}
+
+	public int getBlackValue() {
+		return this.blackValue;
+	}
+	
+	public double getGammaValue() {
+		return this.gammaValue;
+	}
+
+}
